feat(nintendo): product async function implement
