+++ src/controllers/nintendo/productController.js
@@ -2,7 +2,6 @@ const { getProducts } = require("../../services/nintendo/product");
 
 async function getProductList(req, res) {
   try {
-    console.log("getproducts");
     const products = await getProducts();
     res.json({ products });
   } catch (error) {
